+++ src/validataclass/exceptions/__init__.py
@@ -17,7 +17,7 @@ from .dict_exceptions import DictFieldsValidationError, DictInvalidKeyTypeError,
 from .email_exceptions import InvalidEmailError
 from .list_exceptions import ListItemsValidationError, ListLengthError
 from .misc_exceptions import ValueNotAllowedError
-from .number_exceptions import NumberRangeError, DecimalPlacesError, InvalidDecimalError, NonFiniteNumberError
+from .number_exceptions import NumberRangeError, DecimalPlacesError, InvalidIntegerError, InvalidDecimalError, NonFiniteNumberError
 from .regex_exceptions import RegexMatchError
 from .string_exceptions import StringInvalidLengthError, StringTooShortError, StringTooLongError, StringInvalidCharactersError
 from .url_exceptions import InvalidUrlError