+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Official support for Python 3.10.
+- `IntegerValidator`: Add optional boolean parameter `allow_strings` to accept integers as strings (e.g. `"123"`).
 
 ### Fixed